requirements.txt
absl-py==0.1.13
anaconda-client==1.6.3
appdirs==1.4.3
appnope==0.1.0
args==0.1.0
astor==0.6.2
backports.shutil-get-terminal-size==1.0.0
bleach==1.5.0
certifi==2018.10.15
chardet==3.0.4
chart-studio==1.0.0
click==6.7
clint==0.5.1
cloudpickle==0.5.3
clyent==1.2.2
colorlog==2.10.0
cycler==0.10.0
dash==0.30.0
dash-core-components==0.38.0
dash-html-components==0.13.2
dash-renderer==0.15.0
dask==0.18.2
decorator==4.3.0
Django==2.0
django-annoying==0.10.4
django-appconf==1.0.3
django-imagekit==4.0.2
editdistance==0.5.2
entrypoints==0.2.3
Flask==1.0.2
Flask-Compress==1.4.0
floyd-cli==0.10.37
gast==0.2.0
grpcio==1.10.0
gunicorn==19.9.0
h5py==2.7.0
html5lib==0.9999999
idna==2.7
imageio==2.4.1
ipykernel==4.6.1
ipython==6.1.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.10.2
jieba==0.39
Jinja2==2.10
jsonschema==2.6.0
jupyter-client==5.2.3
jupyter-core==4.4.0
jupyterlab==0.34.2
jupyterlab-launcher==0.13.1
Keras==2.1.5
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.0.1
Markdown==2.6.11
MarkupSafe==1.0
marshmallow==2.15.0
matplotlib==2.2.2
mistune==0.7.4
mock==3.0.5
nba-py==0.1.1a2
nbconvert==5.2.1
nbformat==4.4.0
networkx==2.1
nltk==3.3
notebook==5.4.0
numexpr==2.6.6
numpy==1.14.2
olefile==0.44
pandas==0.24.0
pandocfilters==1.4.2
pathlib2==2.3.0
patsy==0.5.1
pexpect==4.2.1
pickleshare==0.7.4
pilkit==2.0
Pillow==5.2.0
pipenv==2018.11.26
plotly==3.4.1
prompt-toolkit==1.0.15
protobuf==3.7.1
ptyprocess==0.5.2
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.6.1
python-graph-core==1.8.2
python-Levenshtein==0.12.0
pytz==2018.7
PyWavelets==0.5.2
PyYAML==3.12
pyzmq==16.0.2
raven==6.6.0
requests==2.20.1
requests-toolbelt==0.8.0
retrying==1.3.3
scikit-image==0.14.0
scikit-learn==0.19.2
scipy==0.19.1
seaborn==0.8
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.11.0
sk-video==1.1.10
sklearn==0.0
statsmodels==0.10.1
tables==3.4.4
tabulate==0.8.2
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.3
toolz==0.9.0
tornado==4.5.2
tqdm==4.25.0
traitlets==4.3.2
urllib3==1.24.1
virtualenv==16.7.5
virtualenv-clone==0.5.3
wcwidth==0.1.7
Werkzeug==0.14.1
whitenoise==4.1.3
xlrd==1.1.0
xlwt==1.3.0
